Add new column in table in analyse schema
Add a new boolean column, named dq_ok (meaning "data quality: OK") to your table in the final schema that holds the aggregated prices. Column should have True if and only if the new data quality requirement is true for that given row (at least 5 different companies and 2 different suppliers provide data for that day and lane and equipment type).